feat: remove create screen (#380)
